backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
sqlalchemy
python-dotenv
pydantic[email]
requests
fastapi-mail
jinja2
